+++ Assets/Packages/com.studio23.ss2.achievementsystem/Editor/Installer.cs
@@ -1,0 +1,15 @@
+using UnityEditor;
+using UnityEngine;
+
+namespace Studio23.SS2.AchievementSystem.Editor
+{
+    public class Installer : UnityEditor.Editor
+    {
+        [MenuItem("Studio-23/AchievementSystem/Install")]
+        public static void Install()
+        {
+           GameObject achievementSystem=new GameObject("Achievement System");
+            achievementSystem.AddComponent<Core.AchievementSystem>();
+        }
+    }
+}