+++ Assets/Packages/com.studio23.ss2.achievementsystem/Runtime/Data/Enums.cs
@@ -1,0 +1,12 @@
+namespace Studio23.SS2.AchievementSystem.Data
+{
+    public enum PlatformProvider
+    {
+        Default,
+        Steam,
+        XBoxCore,
+        XBoxPC,
+        PlayStation,
+        Nintendo
+    }
+}